backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Data Processing
pandas==2.1.3
numpy==1.26.3
scikit-learn==1.3.2
scipy==1.11.4

# API & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# ML & Analytics
joblib==1.3.2
statsmodels==0.14.0

# Utilities
requests==2.31.0
python-multipart==0.0.6
